Latest DDL with Kronos updates too

+++ install/sql/zz_crest_killmail.sql
@@ -3,7 +3,7 @@ DROP TABLE IF EXISTS `zz_crest_killmail`;
 CREATE TABLE `zz_crest_killmail` (
   `killID` int(16) NOT NULL,
   `hash` varchar(64) NOT NULL,
-  `processed` tinyint(1) NOT NULL DEFAULT '0',
+  `processed` smallint(1) NOT NULL DEFAULT '0',
   PRIMARY KEY (`killID`,`hash`),
   KEY `killID` (`killID`),
   KEY `hash` (`hash`),
